refactorig dto's
