+++ controllers/user.ts
@@ -1,8 +1,5 @@
-import type ICustomRequest from '../dto/request/ICustomRequest'
-import type IUserRequest from '../dto/user/IUserRequest'
-import type IUserResponse from '../dto/user/IUSerResponse'
+import { type ICustomRequest, type IUserRequest, type IUserResponse, Status } from '../dto'
 import { type Request, type Response } from 'express'
-import { Status } from '../dto/enums/status'
 import { User } from '../models'
 import { createUserInstanceService, createUserService, getUserbyIdService } from '../services/user'
 import { userRepository } from '../repositories'
