+++ dto/enums/status.ts
@@ -3,3 +3,5 @@ export enum Status {
   INACTIVE = 2,
   BANNED = 3
 }
+
+export default Status
