+++ dto/index.ts
@@ -1,0 +1,21 @@
+import type ICustomRequest from './request/ICustomRequest'
+import type IFilter from './filter/IFilterUser'
+import type IJwtPayload from './auth/IJwtpayload'
+import type ILoginRequest from './auth/ILoginRequest'
+import type IMovieRequest from './movie/IMovieRequest'
+import type IUserRequest from './user/IUserRequest'
+import type IUserResponse from './user/IUserResponse'
+import Roles from './enums/roles'
+import Status from './enums/status'
+
+export {
+  type ICustomRequest,
+  type IFilter,
+  type IJwtPayload,
+  type ILoginRequest,
+  type IMovieRequest,
+  type IUserRequest,
+  type IUserResponse,
+  Roles,
+  Status
+}
