+++ services/auth/index.ts
@@ -1,8 +1,8 @@
 
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
-import type IJwtPayload from '../../dto/auth/IJwtpayload'
 import errorMessageHandler from '../errorMessage'
+import { type IJwtPayload } from '../../dto'
 
 const encrypt = async (text: string): Promise<string> => {
   try {
