+++ services/user/index.ts
@@ -1,8 +1,8 @@
 import { body } from 'express-validator'
-import type IUserRequest from '../../dto/user/IUserRequest'
 import { type Role, type User } from '../../models'
 import { userRepository } from '../../repositories'
 import { encrypt } from '../auth'
+import { type IUserRequest } from '../../dto'
 
 const existsUserById = async (id: number): Promise<void> => {
   const user = await getUserbyIdService(id)
